reorganized for use with CLI service
